Sensor Simulator
The script "SensorSimulator" has been splitted in 3 scripts (or for each sensor), than "Device_Connector" has been created and is the only one that has to be runned. I created just 3 cases, but it can be usefull to create file text with the values of the situation that we want to simulate.